Add reset() method to fix state-save
This is an update on the base class found in
pyswarms/base/base.py. This aims to fix the minor
inconvenience of saving the state of the positions,
velocity, and etc even after the iteration is finished.
In this case, when the optimizer is run once again, it
continues from that state.

The reset() method fixes this by having some attributes
be initialized inside it. In addition, the __init__()
method also calls on reset(), which initializes the said
attributes.

Author: ljvmiranda921