Modify auto-imports in init file
Wow, I hate Python's confusing imports. This is an
update to make sure that the example in README.rst
is working as expected. The init file now can import
some modules, so you can just call the classes as
expected. This should make everything easier. But going
there is quite hard. Probably need to write a tutorial
out of this...

Author: ljvmiranda921