Implemented unit tests for global-best PSO
From this point forward, we'll be implementing a test-driven
development scheme. This is also helpful so that we can be
strict with ourselves come integration time with TravisCI.

For the global best PSO, various checks on edge cases are
currently implemented. Most of these are implemented using
the unittest module. Check `test_gb.py` as a guide.

Author: ljvmiranda921