Implement local best PSO (#3)
This is a feature update where the local-best PSO
was implemented. In lieu of this, the following files
were added to the library:
  - pyswarms/single/lb.py
  - tests/test/optimizers/test_lb.py
  - pyswarms/single/__init__.py

Various checks are also made to the LBestPSO class
as done with GBestPSO. The only difference is that there is
an added check to the number of neighbors and the required
p-value.

Either way, this is all good. Just to please our "audience,"
a soft release of v.0.1.1.dev was done. Congrats!

Author: ljvmiranda921