Improve documentation (#16)
This is a massive documentation update in order to make the
ReadTheDocs website as appealing it should be. In summary, the
following were undertaken:
  - Small patch fix for importing modules
  - Small modification in console_utils for console output
  - Modified ReadTheDocs' sidebar theme with an overriding css file.
  - Added a use-case example on Basic Optimization
  - Updated docstrings for most module implementations to make
    it look nice in the API documentation.

Even though most of these updates occured at different files,
they were all done for the sole purpose of improving documentation
in ReadTheDocs and introducing an API documentation at the end.

Author: ljvmiranda921