Update example on training a neural network (#16)
This is an update to the previous commit regarding the documentation.
A use-case example for training a neural network is now implemented
as a Jupyter Notebook and as an .reST file in ReadTheDocs.

Also, a small stylistic change was done in asserting membership, if
we are to check if a certain key is in a dictionary, we now do the
following

  >>> if <set-of-keys> not in <required-set-of-keys>

Before, we are doing a `if not <set-of-keys> in`, which is
grammatically awkward. The new call sounds more natural and idiomatic.

Author: ljvmiranda921