Added Binary PSO (#7)
Another optimizer, BinaryPSO, has been added. This update
commit introduces test cases in tests/optimizers/test_bn.py
and an abstract discrete class pyswarms/base/dbs.py.

A list of new files can be seen below:
  [Documentation]
  - docs/api/pyswarms.discrete.rst
  [Base Class and Optimizers]
  - pyswarms/base/dbs.py
  - pyswarms/discrete/__init__.py
  - pyswarms/discrete/bn.py
  [Test Cases]
  - tests/optimizers/test_bn.py

This also introduces a small patch for the other base classes.
In this case, the k and p arguments were then transformed into
keyword arguments. They should now be part of the `**kwargs`
dictionary from here on in. This is consistent with the local
best PSO and the binary PSO.

Author: ljvmiranda921