Update docs on feature subset selection (#17)
The documentation was updated to include feature subset
selection. Some examples were also edited from the
pyswarms.single.gb and pyswarms.single.lb modules in
order to fix the error reported from Reddit.

The images for the Use-Case example were then placed
in the same directory as the .reST files.

Author: ljvmiranda921