Add getters for optimization history
The history during optimization is now implemented. Users can
access the cost, position, and velocity histories after
optimization. This works as an interface when implementing the
plotters environment later on.

In this commit, the argument being passed into the base method
_populate_history() is now changed intro a namedtuple. In this
way, implementing new techniques will hopefully be much more
idiomatic because there are keys that are associated for each
history.

For now, this implementation is still naive. This might be
inefficient when dealing with very-large matrices. It may be
better, in the future, to allow state-saving in HDF5 files.
But this is still quite far from the horizon.

Author: ljvmiranda921