Add BaseSearch and RandomSearch
This commit adds a BaseSearch class for GridSearch and RandomSearch,
which defines an __init__ that sets all attributes for instances
as well as score and search methods.

This commit also adds RandomSearch, which provides search for optimal
performance on objective function over combinations of hyperparameter
values with specified bounds for specified number of combination
iterations.

This commit includes 'xrange' functionality for RandomSearch to support
Python 2.7

Author: SioKCronin
Email: siobhankcronin@gmail.com