Fix super() compatibility
This commit fixes the super() compatibility
between Python 2 and 3 versions. The problem
was resolved by passing the class name and
an instance of self to the super() as arguments.

TODO: There should be an instantiation
checker for the search methods. When calling
assertions() from base., please use
super(ClassName, self).assertions()

Author: ljvmiranda921
Email: ljvmiranda921@gmail.com