Add direct import to search tools
This commit updates the __init__.py file to include the two
search methods when importing the pyswarms.utils.search
module. In this case, we don't have to include grid_search
and random_search explicitly.

In addition, I removed the comments in the environments
initialization file to avoid confusion as to where to put
the search methods.

Author: ljvmiranda921
Email: ljvmiranda921@gmail.com