Add test_plot_environment dependencies (#30)
This commit aims to resolve all dependency issues in the
matplotlib.animation.FuncAnimation when running the
test_plot_environment script.

A toolchain called libav-tools is installed in travis-ci,
and an if-else script checks if display is found. If not,
then it uses the default 'Agg' backend.

Moreover, the mock==2.0.0 package is also used during
tests as shown in setup.py.

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com