Add backend module
Reference: #111

This commit adds a backend module consisting of various abstractions
to common PSO operations. There are two files for this:
  - generators: for generating swarm positions, velocities, etc.
  - operators: for various swarm operations (update velocity, etc.)

As of now, one gripe I have with my current implementation is that
it consists of methods with a lot of parameters (it doesn't look
pleasing or concise). My idea is to add a Swarm class that contains
the position, velocity, c1, c2, w, etc. And just put them inside
the optimizer. We'll see how it goes

Signed-off-by: ljvmiranda921 <lester.miranda@obf.ateneo.edu>