Use README.md instead of rst file
In this commit, we'll use the Markdown file going forward. Turns
out the PyPI is now more accommodating of .md files so we'll follow
them from now on.

In the same vein, we also updated the references found in the setup.py
and MANIFEST.in of our module. In py34, there are some problems when
we're not referencing properly in MANIFEST.in, and I wonder why.

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>