Add DeprecationWarning to PlotEnvironment module
In this commit, we add a DeprecationWarning for the PlotEnvironment
module. Although useful, I realized that it might be a hassle when the
environment still optimizes the optimizer if we just want to plot it.

What if a swarm iteration takes a long time? Will plotting take the same
amount of time? We want to decouple the optimization part from the
visualization.

It is better to just have the user supply the position and
cost histories and we just have the module do what it knows.

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>