Add static/dynamic functionality for topologies
Added two new attributes to the topologies. "static" decides whether the topology is a static or a dynamic one, it is initialized together with the class. "neighbor_idx" is an array that stores the indices of the neighbors of every particle. Changed all occurrences of topologies to fit the new initialization. Reworked the tests to fit the new functionality and added more parametrization of fixture functions. Updated the documentation to incorporate the new functionality.

See also: #129