Add plotter functionalities (#172)
Resolves #167 

1. The designer class in formatters.py is used for 3D graphing, but the default label 
list attribute only includes two axes. This causes an index out of range error. 
I added "z-axis".

2. The delta attribute of the Mesher class in formatters.py is 0.001. This is too detailed 
and can lead to mesher objects that are large enough to cause memory errors. 
I changed it to 0.1.

3. The user can't choose a colormap when graphing, so I added a colormaps property 
to the Designer class in formatters.py to allow the user to choose whichever one they 
want. It uses the same style as other properties of the class, and the module imports 
matplotlib.cm to validate both types of colormaps in matplotlib. 

I edited plotters/plot_countour to initialize a Designer object with a default colormap 
if none is provided. I've tested with every category of colormap as well as custom 
colormaps and verified that it is working as intended.

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>
Commited-with: Jay Speidell (@jayspeidell)