Add objective functions (#168)
Reference: #158 

Verified Goldstein and Booth functions working and removed TODO tags. Fixed Bulkin No 6 and removed TODO tag. Implemented Cross-in-Tray, Easom, Eggholder, Himmelblau's, Holder Table, and Three Hump Camel functions. The file is getting long so I sorted the functions and added a list of functions and their corresponding function name to the docstring.
Added the new functions in the test files for the objective functions.

Signed-off-by: Aaron (@whzup)
Commited-with: Jay Speidell (@jayspeidell)