Fix bug in topology classes (#176)
Resolves #170

Fixed a bug in the topology class where not the best values were used as result when optimizing with the LocalBest class. Made the tests for the topologies less trivial by changing the index of the best cost and best position in the fixture swarm. Added a bigger fixture swarm to test the topology classes. Additionally, I fixed a little bug in the Random topology where the elements were compared to 0 instead of np.inf. Deleted the fixture for the neighbour number and replaced it with a pytest.mark.parametrize().

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>
Committed-with: Aaron (@whzup)