Fix bug in the Pyramid topology
Fixed a bug in the Pyramid topology where the optimize function of the GeneralOptimizer would output a ValueError because for some particles the neighbours were not computed. This is achieved by passing custom qhull_options to the Delaunay class. For these qhull_options the necessary documentation and a link to the qhull website was added.
Additionally, changed the boundary for the connectivity of the particles in the Pyramid topology to be swarm.dimensions + 1 instead of 5. The reason for this change is that a n-dimensional simplex has n+1 vertices.

Resolves #183