Clean up changes for v.0.3.0 (#189)
Resolves #183 

Set the Star topology to static and only assign the neighbor_idx when it has not 
already been assigned.

Added the VonNeumann topology to the GeneralOptimizer with the necessary 
changes to the documentation and new exceptions for missing attributes.

Added the VonNeumann topology to the GeneralOptimizer tests. The exceptions 
are tested and the topology fixture now also uses the VonNeumann class. 
Cleaned up the tests and added a static parametrization.

Fixed a bug in the Pyramid topology where the optimize function of the 
GeneralOptimizer would output a ValueError because for some particles 
the neighbours were not computed. This is achieved by passing custom 
qhull_options to the Delaunay class. For these qhull_options the necessary 
documentation and a link to the qhull website was added.

Additionally, changed the boundary for the connectivity of the particles in 
the Pyramid topology to be swarm.dimensions + 1 instead of 5. The reason 
or this change is that a n-dimensional simplex has n+1 vertices.


Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>
Committed-by: Aaron Moser (@whzup )