Update documentation
Updated the documentation of the base class to include the
GeneralOptimizer. Additionally, changed the chapter "Writing unit test"
in the dev.optimizer.rst because we don't use the unittest module
anymore. I replaced the "unittest" module with the "pytest" module and
rewrote some parts of the paragraph so it reads a bit more easily. Also,
added a link to the pytest docs for the running of tests.