Fix bug in topologies (#253)
Fixed a bug in the topologies where all the topologie would calculate
the global best position instead of the best positions of the
neighbourhoods.

Fixed the VonNeumann tests as well as the optimizers such that they
return the position with the best cost. I had to change the parameters
for the optimization in the test_objective_func_with_kwargs to satisfy
the conditions.

Adjusted the code to fit the refactoring of the code. Especially in the
topology tests where the sytax was wrong.

Fixed the test by just checking if either the best or the second best
position is inside of the matrix of the best positions...

Fixed the error with the class passing by adding parenthesis which seem
to have disappeared

Cleaned up the PR and skipped the tests which require a very good cost
as some topologies do not converge fast enough.

Resolves #250